+++ vmm/core/token/handle.go
@@ -2,7 +2,6 @@ package token
 
 import (
 	"encoding/json"
-	"errors"
 	"fmt"
 	"maps"
 	"math/big"
@@ -15,50 +14,11 @@ import (
 	goarSchema "github.com/permadao/goar/schema"
 )
 
-func (h *Token) handleCacheInitial() (res vmmSchema.Result, err error) {
-	if !h.db.CacheInitial() {
-		err = errors.New("cache_initialed")
-		return
-	}
-
-	balances, err := h.db.Balances()
-	if err != nil {
-		return
-	}
-	cacheBalanceMap := make(map[string]string)
-	for k, vl := range balances {
-		if vl == nil {
-			vl = big.NewInt(0)
-		}
-		cacheBalanceMap["balances:"+k] = vl.String()
-	}
-
-	stakes, err := h.db.Stakes()
-	if err != nil {
-		return
-	}
-	cacheStakeMap := make(map[string]string)
-	for k, vl := range stakes {
-		if vl == nil {
-			vl = big.NewInt(0)
-		}
-		cacheStakeMap["stakes:"+k] = vl.String()
-	}
-
-	res.Cache = map[string]string{}
-	maps.Copy(res.Cache, cacheBalanceMap)
-	maps.Copy(res.Cache, cacheStakeMap)
-	maps.Copy(res.Cache, h.cacheTotalSupply())
-	maps.Copy(res.Cache, h.cacheTokenInfo())
-
-	h.db.CacheInitialed()
-	return
-}
-
 func (h *Token) handleInfo(from string) (res vmmSchema.Result, err error) {
 	info := h.Info()
 	decimals := fmt.Sprintf("%v", info.Decimals)
 
+	res.Cache = h.initCache()
 	res.Messages = []*vmmSchema.ResMessage{
 		{
 			Target: from,
@@ -337,50 +297,4 @@ func (h *Token) parseAndCheckAmount(qty string) (*big.Int, error) {
 		return nil, schema.ErrNegativeQuantity
 	}
 	return amt, nil
-}
-
-func (h *Token) cacheTokenInfo() map[string]string {
-	info := h.Info()
-	tokenInfo := map[string]string{
-		"Name":         info.Name,
-		"Ticker":       info.Ticker,
-		"Logo":         info.Logo,
-		"Denomination": fmt.Sprintf("%d", info.Decimals),
-		"Decimals":     fmt.Sprintf("%d", info.Decimals),
-	}
-	res, _ := json.Marshal(tokenInfo)
-	return map[string]string{
-		"token-info": string(res),
-	}
-}
-
-func (h *Token) cacheChangeBalance(updateAccounts ...string) map[string]string {
-	cacheMap := make(map[string]string)
-	for _, accId := range updateAccounts {
-		bal, err := h.db.BalanceOf(accId)
-		if err != nil {
-			bal = big.NewInt(0)
-		}
-		cacheMap["balances:"+accId] = bal.String()
-	}
-	return cacheMap
-}
-
-func (h *Token) cacheTotalSupply() map[string]string {
-	cacheMap := make(map[string]string)
-	totalSupply, err := h.db.GetTotalSupply()
-	if err == nil {
-		cacheMap["total-supply"] = totalSupply.String()
-	}
-	return cacheMap
-}
-
-func (h *Token) cacheChangeStake(accId string) map[string]string {
-	cacheMap := make(map[string]string)
-	stake, err := h.db.StakeOf(accId)
-	if err != nil {
-		stake = big.NewInt(0)
-	}
-	cacheMap["stakes:"+accId] = stake.String()
-	return cacheMap
 }
